+++ packages/style-value-types/src/index.ts
@@ -1,25 +1,15 @@
 // Types
-import { Transformer, ValueType, NumberMap, RGBA, HSLA, Color } from './types';
-export { Transformer, ValueType, NumberMap, RGBA, HSLA, Color };
+export { Transformer, ValueType, NumberMap, RGBA, HSLA, Color } from './types';
 
 // Value types
-import { number, scale, alpha } from './value-types/numbers';
-import { degrees, percent, px, vw, vh } from './value-types/units';
-import { rgba, rgbUnit, hsla, hex, color } from './value-types/color';
-import complex from './value-types/complex';
+export { number, scale, alpha } from './value-types/numbers';
 export {
-  number,
-  scale,
-  alpha,
   degrees,
   percent,
+  progressPercentage,
   px,
   vw,
-  vh,
-  rgba,
-  rgbUnit,
-  hex,
-  hsla,
-  color,
-  complex
-};
+  vh
+} from './value-types/units';
+export { rgba, rgbUnit, hsla, hex, color } from './value-types/color';
+export { complex } from './value-types/complex';